+++ opendc-trace/opendc-trace-opendc/src/main/kotlin/org/opendc/trace/opendc/OdcVmResourceTableReader.kt
@@ -25,6 +25,7 @@ package org.opendc.trace.opendc
 import org.apache.avro.Schema
 import org.apache.avro.generic.GenericRecord
 import org.opendc.trace.*
+import org.opendc.trace.conv.*
 import org.opendc.trace.util.parquet.LocalParquetReader
 import java.time.Instant
