+++ opendc-trace/opendc-trace-swf/src/main/kotlin/org/opendc/trace/swf/SwfTaskTableReader.kt
@@ -23,6 +23,7 @@
 package org.opendc.trace.swf
 
 import org.opendc.trace.*
+import org.opendc.trace.conv.*
 import java.io.BufferedReader
 import java.time.Duration
 import java.time.Instant
